newsfeed hbase
