+++ django_hbase/models/exceptions.py
@@ -1,0 +1,6 @@
+class BadRowKeyError(Exception):
+    pass
+
+
+class EmptyColumnError(Exception):
+    pass
