+++ django_hbase/models/hbase_models.py
@@ -1,14 +1,7 @@
 from django_hbase.models import HBaseField, IntegerField, TimestampField
 from django_hbase.client import HBaseClient
 from django.conf import settings
-
-
-class BadRowKeyError(Exception):
-    pass
-
-
-class EmptyColumnError(Exception):
-    pass
+from django_hbase.models.exceptions import BadRowKeyError, EmptyColumnError
 
 
 class HBaseModel:
@@ -191,14 +184,22 @@ class HBaseModel:
             row_data[column_key] = cls.serialize_field(field, column_value)
         return row_data
 
-    def save(self):
+    def save(self, batch=None):
         row_data = self.serialize_row_data(self.__dict__)
         # 如果 row_data 为空，即没有任何 column key values 需要存储 hbase 会直接不存储
         # 这个 row_key, 因此我们可以 raise 一个 exception 提醒调用者，避免存储空值
         if len(row_data) == 0:
             raise EmptyColumnError
-        table = self.get_table()
-        table.put(self.row_key, row_data)
+        if batch:
+            # batch.put()不会立刻产生一个HBase的数据库请求
+            # 而是会等到batch.send()执行后，一次性把所有的批量数据都写入到HBase
+            # batch的好处是：通常情况下HBase和Web Server不在同一台机器上，这样每次写
+            # HBase都会产生一个类似request的请求，而batch只会产生一次请求，
+            # 可以节省时间
+            batch.put(self.row_key, row_data)
+        else:
+            table = self.get_table()
+            table.put(self.row_key, row_data)
 
     @classmethod
     def get(cls, **kwargs):
@@ -258,10 +259,20 @@ class HBaseModel:
 
     # step 1：创建HBaseModel
     @classmethod
-    def create(cls, **kwargs):
+    def create(cls, batch=None, **kwargs):
         # 类似django ORM的写法
         # 调用__init__构造函数
         instance = cls(**kwargs)
-        instance.save()
+        instance.save(batch=batch)
         return instance
+
+    @classmethod
+    def batch_create(cls, batch_data):
+        table = cls.get_table()
+        batch = table.batch()
+        results = []
+        for data in batch_data:
+            results.append(cls.create(batch=batch, **data))
+        batch.send()
+        return results
