+++ friendships/models/__init__.py
@@ -1,0 +1,2 @@
+from .friendship import *
+from .hbase_friendship import *
