+++ friendships/models/hbase_friendship.py
