+++ friendships/tests.py
@@ -1,8 +1,7 @@
-from friendships.models import Friendship
 from friendships.services import FriendshipService
 from testing.testcases import TestCase
 from django_hbase.models import EmptyColumnError, BadRowKeyError
-from friendships.hbase_models import HBaseFollower, HBaseFollowing
+from friendships.models import HBaseFollower, HBaseFollowing
 
 import time
