+++ newsfeeds/models/__init__.py
@@ -1,0 +1,2 @@
+from .hbase_newsfeed import *
+from .newsfeed import *
