+++ tweets/models/__init__.py
@@ -1,0 +1,2 @@
+from .tweet import Tweet
+from .tweet_photo import TweetPhoto
